OpencvProject/requirements.txt
asgiref==3.5.2
click==8.1.3
cmake==3.24.1
colorama==0.4.5
cycler==0.11.0
Django==4.1
dlib @ file:///C:/Users/DIMOSO%20EL%20JUNIOR/Desktop/OPEN/OpencvProject/OpencvFiles/dlib-19.22.99-cp39-cp39-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.36.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
